SpatioTemporalTraits

SpatioTemporalTraits.assert_timedim_last — Method

assert_timedim_last(x)

Throw an error if the x has a time dimension that is not the last dimension.

SpatioTemporalTraits.depth — Method

depth(x)

Returns the size of the dimension corresponding to depth.

SpatioTemporalTraits.duration — Method

duration(x)

Duration of the event along the time axis.

SpatioTemporalTraits.has_timedim — Method

has_timedim(x) -> Bool

Returns 'true' if x has a time dimension.

SpatioTemporalTraits.height — Method

height(x)

Returns the size of the dimension corresponding to height.

SpatioTemporalTraits.is_spatial — Method

is_spatial(x::StaticSymbol) -> StaticBool

Returns static(true) if x refers to a spatial dimension. By default, any dimension that is not one of the following is considered spatial:

time, Time

channels, Channels

observations, Observations, obs

SpatioTemporalTraits.is_temporal — Method

is_temporal(x) -> StaticBool

Returns True if x refers to time.

SpatioTemporalTraits.ntimes — Method

ntimes(x) -> Integer

Returns the number of elements along the time dimension.

SpatioTemporalTraits.origin — Method

origin(x) -> Tuple

Returns the spatial origin of x.

SpatioTemporalTraits.pixel_spacing — Method

pixel_spacing(x)

Return a tuple representing the separation between adjacent pixels along each axis of the image. Derived from the step size of each element of spatial_indices.

SpatioTemporalTraits.sampling_rate — Method

sampling_rate(x)

Number of samples per second.

SpatioTemporalTraits.sdims — Method

sdims(x) -> Int

Return the number of spatial dimensions that x has.

See also: spatialdims, is_spatial

SpatioTemporalTraits.spatial_directions — Method

spatial_directions(img) -> (axis1, axis2, ...)

Return a tuple-of-tuples, each axis[i] representing the displacement vector between adjacent pixels along spatial axis i of the image array, relative to some external coordinate system ("physical coordinates").

By default this is computed from pixel_spacing, but you can set this manually using ImageMeta.

SpatioTemporalTraits.spatial_first — Method

spatial_first(x)

The first position along each spatial indices (i.e., where each spatial axis starts).

SpatioTemporalTraits.spatial_indices — Method

spatial_indices(x)

Returns the keys along each spatial dimension.

SpatioTemporalTraits.spatial_last — Method

spatial_last(x)

The last position along each spatial indices (i.e., where each spatial axis stops).

SpatioTemporalTraits.spatial_order — Method

spatial_order(x) -> Tuple

Returns the dimnames of x that correspond to spatial dimensions, as determined by is_spatial. At most three dimensions will be considered spatial.

SpatioTemporalTraits.spatialdims — Method

spatialdims(x) -> Tuple{Vararg{StaticInt}}

Return a tuple of the spatial dimensions of x.

SpatioTemporalTraits.time_first — Method

time_first(x)

First time point along the time axis.

SpatioTemporalTraits.time_last — Method

time_last(x)

Last time point along the time axis.

SpatioTemporalTraits.time_step — Method

time_step(x)

The time step/interval between each element.

SpatioTemporalTraits.timedim — Method

timedim(::Type{T}) -> Integer

Returns the dimension associated with time.

SpatioTemporalTraits.times — Method

times(x)

Returns a collection of times associated with x. If x has a specific dimension associated with time, this returns the keys along that dimension.

SpatioTemporalTraits.width — Method

width(x)

Returns the size of the dimension corresponding to width.